Support -std=f2018
The Fortran committee has decided to rename the upcoming Fortran 2015
standard to Fortran 2018.  This is not a reflection of a three year
delay in the process, but rather they are following other standards in
adopting the year of publication for the name. For more details see
N2144.

This patch renames GFC_STD_F2015 to GFC_STD_F2018, and makes it a
separate flag rather than an alias for GFC_STD_GNU. Also, it adds a
-std=f2018 argument, and documents it.

Regtested on x86_64-pc-linux-gnu.

gcc/fortran/ChangeLog:

2017-12-17  Janne Blomqvist  <jb@gcc.gnu.org>

	* decl.c (gfc_match_implicit_none): Use GFC_STD_F2018 instead of
	GFC_STD_F2015.
	* error.c (gfc_notify_std): Add GFC_STD_F2018{_DEL,_OBS} to
	switch.
	* gfortran.texi: Document -std=f2018.
	* interface.c (compare_parameter): Fix comment.
	* invoke.texi: Document -std=f2018.
	* lang.opt: Add -std=f2018 argumnet.
	* libgfortran.h (GFC_STD_F2015): Rename to GFC_STD_F0218, use
	separate flag bit.
	(GFC_STD_F2018_DEL): New macro.
	(GFC_STD_F2018_OBS): Likewise.
	* match.c (gfc_match_stopcode): Use GFC_STD_F2018.
	* options.c (set_default_std_flags): Add F2018 flags to defaults.
	(gfc_handle_option): Set options for -std=f2018.

gcc/testsuite/ChangeLog:

2017-12-17  Janne Blomqvist  <jb@gcc.gnu.org>

	* gfortran.dg/error_stop_3.f90: Update -std= option, fix comments.
	* gfortran.dg/error_stop_4.f90: Update error message.
	* gfortran.dg/implicit_14.f90: Likewise.
	* gfortran.dg/spellcheck-procedure_2.f90: Don't warn for F2018
	features.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@255761 138bc75d-0d04-0410-961f-82ee72b054a4